# indexes: Add compact block filter headers cache