@theStack 2020-05-13
I'm almost reading for ACKing this PR, just have some minor suggestion/discussion points:
- Would it make sense to also only check the cache if we have checkpoints, i.e. a method structure like
```
bool is_checkpoint = (block_index->nHeight % CFCHECKPT_INTERVAL == 0); 
if (is_checkpoint)
    // Check the headers cache

// Regular lookup

if (is_checkpoint)
    // Add to the headers cache
```
? Of course, the drawback of this approach is that the condition for checking needs to be removed if the cache is ever used for other filter headers than those from the checkpoints. Also not sure if its really worth it from a performance point of view. But in 999 out of 1000 times the lookup in the map could be skipped.
- In the regular case the cache should never get full, considering that this would by plan only happen in the far future (block 2000000, >25 years from now, with the current maximum size constant of 2000). If it gets full however, should we take some action like putting a debug output or even an assert, to get noticed that we either have some memory exhaustion bug or that the maximum size needs to be increased?
- Any idea how to effectively test performance improvement like this? I don't know anything better than just putting some `LogPrintf`s in the cache hit / cache add cases, run the (maybe slightly modified) functional test and analyze the log output to check if it behaves as expected.