@jnewbery 2020-05-13
@theStack 
> Would it make sense to also only check the cache if we have checkpoints, i.e. a method structure like

Done

@MarcoFalke 
> a non-recursive mutex should be sufficient.

Changed `RecursiveMutex` -> `Mutex`

I'm still working on adding a microbench, although it's not clear to me that it'll show any improvement since the disk reads may be cached by the OS for the microbench.