@theStack 2020-05-14
Using jnewbery's nice [node-shell module](https://github.com/jnewbery/bitcoin/tree/node-shell) (a modification of test-shell that allows to connect with the functional test framework to a running full-node) I crafted up a quick and dirty [test-script](https://github.com/theStack/bitcoin/blob/pr18960_test/test/functional/p2p_18960.py) (see branch https://github.com/theStack/bitcoin/tree/pr18960_test) that sends `getcfcheckpt` requests with the stop-hash of the halving block 630000 and measures the processing time.

Freshly started bitcoind from master branch (w/o checkpoint headers cache implementation), after mempool imports are done:
```
getcfcheckpt request processing times
=====================================
-> first:     53.75ms
-> following: 50.70ms (average)
=====================================
```

Freshly started bitcoind from PR branch (with checkpoint headers cache implementation), after mempool imports are done:
```
getcfcheckpt request processing times
=====================================
-> first:     50.92ms
-> following: 50.80ms (average)
=====================================
```
There doesn't seem to be a noticable difference. I don't know if my methodology is flawed though. Maybe all the processing around is taking so much more time that the actual part of fetching the header (from the cache or from disk) is neglectible. If anyone has an idea how to improve the benchmark script, happy to hear suggestions!