@jnewbery 2020-05-15
@narula suggested in today's PR Review Club meeting that it'd be less cognitive overhead to simply take the `m_cs_headers_cache` for the entirety of the `LookupFilterHeader()` function, rather than taking, releasing and taking it again. That seem like a reasonable request to me. We can't ever block important work with this lock, so holding it a bit longer than strictly necessary is not an issue.