@jnewbery 2020-05-15
@theStack - thank you for doing some benching of this. Your approach inspired me to do something similar myself, using debug logs to find the exact time to construct a cfcheckpt response. I've added a [`debug_log_delta` context wrapper](https://github.com/jnewbery/bitcoin/commit/99ea8d5b228c3c5ddec5de76bbf830c666738485#diff-86294e5ae5283eebdd9f98d79007a0e1R386) to the node_shell, which allows me to provide two debug log messages, then do actions on the node and extract the time delta between those two logs. I applied that patch to:

- without a cache
- with this cache
- with @narula's suggestion of taking the lock only once (I was interested to see if there were any performance differences).

And then ran the following node-shell script:

```
REPO_PATH="<path_to_repo>"
DATADIR_PATH="<path_to_datadir>"
import sys
sys.path.insert(0, f"{REPO_PATH}/test/functional")
from test_framework.node_shell import NodeShell
test = NodeShell()
test.setup(datadir=DATADIR_PATH)
# <test_framework.node_shell.NodeShell.__TestShell object at 0x7f7704820490>
node = test.nodes[0]
bb = node.getbestblockhash()
from test_framework.messages import FILTER_TYPE_BASIC, msg_getcfcheckpt
request = msg_getcfcheckpt(filter_type=FILTER_TYPE_BASIC, stop_hash=int(bb, 16))
for i in range(15):
    with node.debug_log_delta("getcfcheckpt request received", "cfcheckpt response constructed"):
        node.p2ps[0].send_and_ping(request)
test.shutdown()
```

### [no cache](https://github.com/jnewbery/bitcoin/tree/2020-05-nocfcheckpts-cache-debug)

```
getcfcheckpt request received to cfcheckpt response constructed: 7.824ms
getcfcheckpt request received to cfcheckpt response constructed: 7.054ms
getcfcheckpt request received to cfcheckpt response constructed: 7.021ms
getcfcheckpt request received to cfcheckpt response constructed: 7.303ms
getcfcheckpt request received to cfcheckpt response constructed: 7.355ms
getcfcheckpt request received to cfcheckpt response constructed: 7.26ms
getcfcheckpt request received to cfcheckpt response constructed: 7.169ms
getcfcheckpt request received to cfcheckpt response constructed: 7.166ms
getcfcheckpt request received to cfcheckpt response constructed: 7.194ms
getcfcheckpt request received to cfcheckpt response constructed: 8.574ms
getcfcheckpt request received to cfcheckpt response constructed: 7.119ms
getcfcheckpt request received to cfcheckpt response constructed: 7.156ms
getcfcheckpt request received to cfcheckpt response constructed: 4.153ms
getcfcheckpt request received to cfcheckpt response constructed: 8.162ms
getcfcheckpt request received to cfcheckpt response constructed: 6.485ms
```

There does seem to be a very slight performance improvement after the first request. Maybe around 5-10% from the mean, but variance is quite high, so it might be noise.

### [cache (taking locks twice)](https://github.com/jnewbery/bitcoin/tree/2020-05-cfcheckpts-cache-debug)

```
getcfcheckpt request received to cfcheckpt response constructed: 12.2ms
getcfcheckpt request received to cfcheckpt response constructed: 2.855ms
getcfcheckpt request received to cfcheckpt response constructed: 2.992ms
getcfcheckpt request received to cfcheckpt response constructed: 2.869ms
getcfcheckpt request received to cfcheckpt response constructed: 3.141ms
getcfcheckpt request received to cfcheckpt response constructed: 3.692ms
getcfcheckpt request received to cfcheckpt response constructed: 3.452ms
getcfcheckpt request received to cfcheckpt response constructed: 2.916ms
getcfcheckpt request received to cfcheckpt response constructed: 3.413ms
getcfcheckpt request received to cfcheckpt response constructed: 3.648ms
getcfcheckpt request received to cfcheckpt response constructed: 3.816ms
getcfcheckpt request received to cfcheckpt response constructed: 3.84ms
getcfcheckpt request received to cfcheckpt response constructed: 3.812ms
getcfcheckpt request received to cfcheckpt response constructed: 3.869ms
getcfcheckpt request received to cfcheckpt response constructed: 3.445ms
```

The initial request is slower (presumably because the cache needs to be warmed), but subsequent requests are on average >50% faster than with no cache.

### [cache (taking lock once)](https://github.com/jnewbery/bitcoin/tree/2020-05-cfcheckpts-onelock-cache-debug)

```
getcfcheckpt request received to cfcheckpt response constructed: 11.785ms
getcfcheckpt request received to cfcheckpt response constructed: 3.35ms
getcfcheckpt request received to cfcheckpt response constructed: 3.19ms
getcfcheckpt request received to cfcheckpt response constructed: 3.475ms
getcfcheckpt request received to cfcheckpt response constructed: 3.175ms
getcfcheckpt request received to cfcheckpt response constructed: 3.345ms
getcfcheckpt request received to cfcheckpt response constructed: 3.274ms
getcfcheckpt request received to cfcheckpt response constructed: 3.161ms
getcfcheckpt request received to cfcheckpt response constructed: 3.259ms
getcfcheckpt request received to cfcheckpt response constructed: 3.217ms
getcfcheckpt request received to cfcheckpt response constructed: 3.238ms
getcfcheckpt request received to cfcheckpt response constructed: 3.253ms
getcfcheckpt request received to cfcheckpt response constructed: 3.33ms
getcfcheckpt request received to cfcheckpt response constructed: 3.32ms
getcfcheckpt request received to cfcheckpt response constructed: 3.36ms
```

The first request is slightly faster than when taking locks twice (although that might be noise). Subsequent requests are the same speed as the 'taking locks twice' branch, which is what we'd expect because locks are only taken once in that branch if the header is found in the cache.

@MarcoFalke - I think these full node manual benches are more meaningful than microbenches, so I'm not planning to add any automated benches to `/src/bench`.