@theStack 2020-05-15
In yesterday's [review club meeting on this PR](https://bitcoincore.reviews/18960.html), it was suggested to put LogPrintf() directly into the network processing code parts to measure the execution time of getcfcheckpt requests. (My previous approach of [measuring the request response time from the test framework](https://github.com/bitcoin/bitcoin/pull/18960#issuecomment-628761578) was not precise enough, as the time also included python overhead and the time for the message handler to wake up.) I did some testing and can roughly confirm the [results from jnewbery](https://github.com/bitcoin/bitcoin/pull/18960#issuecomment-628980196) (see commit https://github.com/theStack/bitcoin/commit/48cff1bb4d27ef0040b18a860a3c897752e3c296):

*without cache*:
```
2020-05-14T20:15:11Z getcfcheckpt processing took 10247us
2020-05-14T20:15:11Z getcfcheckpt processing took 6105us
2020-05-14T20:15:11Z getcfcheckpt processing took 6437us
2020-05-14T20:15:11Z getcfcheckpt processing took 5330us
2020-05-14T20:15:11Z getcfcheckpt processing took 6964us
2020-05-14T20:15:11Z getcfcheckpt processing took 5523us
2020-05-14T20:15:12Z getcfcheckpt processing took 5413us
2020-05-14T20:15:12Z getcfcheckpt processing took 5843us
2020-05-14T20:15:12Z getcfcheckpt processing took 7060us
2020-05-14T20:15:12Z getcfcheckpt processing took 6165us
...
```
*with cache*:
```
2020-05-14T20:21:26Z getcfcheckpt processing took 9801us
2020-05-14T20:21:26Z getcfcheckpt processing took 2619us
2020-05-14T20:21:26Z getcfcheckpt processing took 3338us
2020-05-14T20:21:26Z getcfcheckpt processing took 2446us
2020-05-14T20:21:26Z getcfcheckpt processing took 2560us
2020-05-14T20:21:27Z getcfcheckpt processing took 2331us
2020-05-14T20:21:27Z getcfcheckpt processing took 2703us
2020-05-14T20:21:27Z getcfcheckpt processing took 2614us
2020-05-14T20:21:27Z getcfcheckpt processing took 2598us
2020-05-14T20:21:27Z getcfcheckpt processing took 2444us
...
```
One can see that there's a speedup of at least 2x with the cache.

Note that my approach to get these numbers was slightly different, as I directly measured the time in bitcoind via `GetTimeMicros()` (commit https://github.com/theStack/bitcoin/commit/49f5f3d5444ed6c872dd41be910e2149308b35e7):
```diff
diff --git a/src/net_processing.cpp b/src/net_processing.cpp
index 1df1fab..6e620d9 100644
--- a/src/net_processing.cpp
+++ b/src/net_processing.cpp
@@ -3380,7 +3380,10 @@ bool ProcessMessage(CNode* pfrom, const std::string& msg_type, CDataStream& vRec
     }

     if (msg_type == NetMsgType::GETCFCHECKPT) {
+        int64_t start_time = GetTimeMicros();
         ProcessGetCFCheckPt(pfrom, vRecv, chainparams, connman);
+        int64_t processing_duration = GetTimeMicros() - start_time;
+        LogPrintf("getcfcheckpt processing took %lldus\n", processing_duration);
         return true;
     }
```

but the idea is the same.

Now I'm very curious if the change from `std::map` to `std::unordered_map` shows a noticable difference in performance.

@MarcoFalke: Nice idea to drop the cache between the requests. Unfortunately, on my current development box I don't have the permissions (root rights) to do this :/ if anyone having the possibility, the results would be very interesting.